Add files via upload
The *completed* R script file from Thursday's class. Feel free to compare with your own answers, and to keep for future reference. You do not need to understand everything that is going on in this file from a code perspective: the purpose of the script is to provide you with a model for good practice in terms of structuring your script files and keeping track of your workflow.